HandleCreateText

+++ internal/model/card.go
@@ -3,14 +3,14 @@ package model
 import "time"
 
 type Card struct {
-	ID                int
-	UserID            int
-	MetadataID        int
+	ID                int64
+	UserID            int64
+	MetadataID        int64
 	CardPaymentSystem string
 	CardNumber        string
 	CardHolder        string
 	CardEndDate       time.Time
-	CVC               int
+	CVC               int64
 	CreatedAt         time.Time
 	UpdatedAt         time.Time
 	DeletedAt         time.Time
